sweet alert added for confirmation

+++ src/Pages/Dashboard/UserRow.js
@@ -1,48 +1,67 @@
 import React from 'react';
 import { RiDeleteBin6Line } from 'react-icons/ri';
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 import { toast } from 'react-toastify';
 
 
 const UserRow = ({ index, user, refetch }) => {
-    const { name, email, img, role } = user
+    const { name, email, img, role } = user;
+    const ConfirmSwal = withReactContent(Swal);
+    const DeleteSwal = withReactContent(Swal);
 
     const makeAdmin = () => {
-        fetch(`https://murmuring-retreat-70420.herokuapp.com/users/admin/${email}`, {
-            method: 'PUT',
-            headers: {
-                'authorization': `Bearer ${localStorage.getItem('accessToken')}`
+        ConfirmSwal.fire({
+            title: 'Do you want to make this user an admin ?',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://murmuring-retreat-70420.herokuapp.com/users/admin/${email}`, {
+                    method: 'PUT',
+                    headers: {
+                        'authorization': `Bearer ${localStorage.getItem('accessToken')}`
+                    }
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.modifiedCount > 0) {
+                            toast('Successfully Make as Admin');
+                            refetch()
+                        }
+                        else {
+                            toast("Something went wrong, Please Try again later.")
+                        }
+                    })
             }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.modifiedCount > 0) {
-                    toast('Successfully Make Admin');
-                    refetch()
-                }
-                else {
-                    toast("Something went wrong, Please Try again later.")
-                }
-            })
+
     }
     const deleteUser = () => {
-        fetch(`https://murmuring-retreat-70420.herokuapp.com/users/${email}`, {
-            method: 'DELETE',
-            headers: {
-                'authorization': `bearer ${localStorage.getItem('accessToken')}`
+        DeleteSwal.fire({
+            title: 'Do you want to delete this user ?',
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://murmuring-retreat-70420.herokuapp.com/users/${email}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'authorization': `bearer ${localStorage.getItem('accessToken')}`
+                    }
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.deletedCount > 0) {
+                            toast('Successfully Delete User');
+                            refetch()
+                        }
+                        else {
+                            toast("Something went wrong, Please Try again later.")
+                        }
+                    })
             }
         })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount > 0) {
-                    toast('Successfully Delete User');
-                    refetch()
-                }
-                else {
-                    toast("Something went wrong, Please Try again later.")
-                }
-            })
-            
-
     }
     return (
 
@@ -52,38 +71,10 @@ const UserRow = ({ index, user, refetch }) => {
             <td>{email}</td>
             <td><img className='rounded-full w-10 mask mask-hexagon' src={img || 'https://api.lorem.space/image/face?hash=92310'} alt={user.product} /></td>
             <td>{role || 'Customer'}</td>
-            <td className='text-center'>
-                <button onClick={deleteUser} className='btn btn-xm'><RiDeleteBin6Line/> Delete</button>
-                <button onClick={makeAdmin} className='btn btn-xm'>Make Admin</button>
-
-                {/* <label htmlFor="admin-modal" className="btn btn-xs btn-success mr-2">Make Admin</label> */}
-                {/* <label htmlFor="delete-modal" className="btn btn-xs btn-error ml-2"><RiDeleteBin6Line className='text-sm mx-2' /> Delete User</label> */}
-            </td>
-            {/* <td>
-                <input type="checkbox" id="admin-modal" className="modal-toggle" />
-                <div className="modal modal-bottom sm:modal-middle">
-                    <div className="modal-box">
-                        <h3 className="text-xl py-5">Are you sure want to make this user as admin?</h3>
-                        <div className="modal-action flex justify-evenly">
-                            <label onClick={confirmAdmin} htmlFor="admin-modal" className="btn">Yes</label>
-                            <label htmlFor="admin-modal" className="btn">Cancel</label>
-                        </div>
-                    </div>
-                </div>
-            </ td>
             <td>
-                <input type="checkbox" id="delete-modal" className="modal-toggle" />
-                <div className="modal modal-bottom sm:modal-middle">
-                    <div className="modal-box">
-                        <h3 className="text-xl py-5">Are you sure want to delete this user?</h3>
-                        <div className="modal-action flex justify-evenly">
-                        <label onClick={deleteUser} htmlFor="delete-modal" className="btn">Yes</label>
-                            <label htmlFor="delete-modal" className="btn btn-primary">Cancel</label>
-                        </div>
-                    </div>
-                </div>
-            </td> */}
-            
+                <button onClick={deleteUser} className='btn btn-xs btn-error mx-1'><RiDeleteBin6Line /> Delete</button>
+                <button onClick={makeAdmin} className='btn btn-xs btn-success mx-1'>Make Admin</button>
+            </td>
         </tr>
